refactor: 장소 도착 시 view를 띄우는 함수를 extension으로 분리
